backend/requirements.txt
alembic
annotated-doc
annotated-types
anyio
bcrypt
cffi
click
colorama
fastapi
greenlet
h11
idna
Mako
MarkupSafe
passlib
psycopg2-binary
pycparser
pydantic
pydantic_core
PyJWT
python-dotenv
SQLAlchemy
starlette
typing-inspection
typing_extensions
uvicorn
xlsxwriter
